+++ src/_/index.tests.ts
@@ -8,6 +8,7 @@ const tokens: Record<keyof typeof libTokens, true> = {
   // css.ts
   css: true,
   media: true,
+  str: true,
 
   // scoped.ts
   scoped: true,